Remove references to image/jpg content type
This content type raises a deprecation warning on Rails 7.0

DEPRECATION WARNING: image/jpg is not a valid content type, it should
not be used when creating a blob, and support for it will be removed
in Rails 7.1. If you want to keep supporting this content type past
Rails 7.1, add it to `config.active_storage.variable_content_types`.

https://www.w3.org/Graphics/JPEG/