Support libvips < 8.8
If you try to validate images while using libvips 8.7 every image fails validation with an "is not a valid image" error. This was helpfully investigated by @marckohlbrugge in #170 and seems to be caused because libvips can't report which file types it supports in versions below 8.8.

I've changed the code so it checks to see whether libvips can report the file types. I've assumed that if we have libvips installed and it can't report the supported file types then we should still use libvips. This might mean we raise errors if we give vips a file it does not support but I can't see another (easy) way to support older versions of libvips.